chess: narrow phase/phase_weight to uint8_t for eZ80
phase only holds 0-24, phase_weight values are 0-4 — uint8_t is
the natural fit. Saves 7 bytes (6B array + 1B board_t field) and
removes dead code (negative phase check was impossible with uint8_t).

Benchmark: perf-neutral vs Phase 1, binary -584B vs baseline.

mg/eg kept as int16_t — widening to int was tested but caused +5.6%
make/unmake regression from increased memory traffic.

+++ chess/engine/src/eval.h
@@ -14,7 +14,7 @@ extern const int16_t eg_table[6][64];
 /* Phase weight per piece type (0-based).
    Pawn=0, Knight=1, Bishop=1, Rook=2, Queen=4, King=0.
    Total starting phase = 24 (PHASE_MAX). */
-extern const int16_t phase_weight[6];
+extern const uint8_t phase_weight[6];
 
 #define PHASE_MAX 24
